+++ icd_loader/icd_loader.h
@@ -480,7 +480,7 @@ typedef xe_result_t (__xecall *pfn_xeModuleBuildLogDestroy)(
     );
 typedef xe_result_t (__xecall *pfn_xeModuleBuildLogGetString)(
     xe_module_build_log_handle_t hModuleBuildLog,   ///< [in] handle of the module build log object.
-    uint32_t* pSize,                                ///< [out] size of build log string.
+    size_t* pSize,                                  ///< [out] size of build log string.
     const char** pBuildLog                          ///< [out] pointer to null-terminated string of the log.
     );
 typedef xe_result_t (__xecall *pfn_xeModuleGetNativeBinary)(
